stationary kernels refactoring
