+++ gpjax/kernels/stationary/matern32.py
@@ -14,14 +14,16 @@
 # ==============================================================================
 
 from dataclasses import dataclass
+from typing import Dict, List, Optional
 
 import jax.numpy as jnp
 import tensorflow_probability.substrates.jax.bijectors as tfb
 import tensorflow_probability.substrates.jax.distributions as tfd
 from jaxtyping import Array, Float
 
-from ...base import param_field
+from ...parameters import Softplus, param_field
 from ..base import AbstractKernel
+from ..computations import DenseKernelComputation
 from .utils import build_student_t_distribution, euclidean_distance
 
 
@@ -29,16 +31,13 @@ from .utils import build_student_t_distribution, euclidean_distance
 class Matern32(AbstractKernel):
     """The Matérn kernel with smoothness parameter fixed at 1.5."""
 
-    lengthscale: Float[Array, "D"] = param_field(
-        jnp.array([1.0]), bijector=tfb.Softplus()
-    )
-    variance: Float[Array, "1"] = param_field(jnp.array([1.0]), bijector=tfb.Softplus())
-    name: str = "Matérn32"
+    lengthscale: Float[Array, "D"] = param_field(jnp.array([1.0]), bijector=Softplus)
+    variance: Float[Array, "1"] = param_field(jnp.array([1.0]), bijector=Softplus)
 
     def __call__(
         self,
-        x: Float[Array, "D"],
-        y: Float[Array, "D"],
+        x: Float[Array, "1 D"],
+        y: Float[Array, "1 D"],
     ) -> Float[Array, "1"]:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with
         lengthscale parameter :math:`\\ell` and variance :math:`\\sigma^2`
@@ -47,8 +46,8 @@ class Matern32(AbstractKernel):
             k(x, y) = \\sigma^2 \\exp \\Bigg(1+ \\frac{\\sqrt{3}\\lvert x-y \\rvert}{\\ell^2}  \\Bigg)\\exp\\Bigg(-\\frac{\\sqrt{3}\\lvert x-y\\rvert}{\\ell^2} \\Bigg)
 
         Args:
-            x (Float[Array, "D"]): The left hand argument of the kernel function's call.
-            y (Float[Array, "D"]): The right hand argument of the kernel function's call.
+            x (Float[Array, "1 D"]): The left hand argument of the kernel function's call.
+            y (Float[Array, "1 D"]): The right hand argument of the kernel function's call.
 
         Returns:
             Float[Array, "1"]: The value of :math:`k(x, y)`.
@@ -60,5 +59,5 @@ class Matern32(AbstractKernel):
         return K.squeeze()
 
     @property
-    def spectral_density(self) -> tfd.Distribution:
+    def spectral_density(self):
         return build_student_t_distribution(nu=3)
