+++ gpjax/kernels/stationary/periodic.py
@@ -16,12 +16,14 @@
 from dataclasses import dataclass
 
 import jax.numpy as jnp
-import tensorflow_probability.substrates.jax.bijectors as tfb
-import tensorflow_probability.substrates.jax.distributions as tfd
+from jax.random import KeyArray
 from jaxtyping import Array, Float
 
 from ...base import param_field
 from ..base import AbstractKernel
+
+from dataclasses import dataclass
+from ...parameters import param_field, Softplus
 
 
 @dataclass
@@ -31,24 +33,23 @@ class Periodic(AbstractKernel):
     Key reference is MacKay 1998 - "Introduction to Gaussian processes".
     """
 
-    lengthscale: Float[Array, "D"] = param_field(
-        jnp.array([1.0]), bijector=tfb.Softplus()
-    )
-    variance: Float[Array, "1"] = param_field(jnp.array([1.0]), bijector=tfb.Softplus())
-    period: Float[Array, "1"] = param_field(jnp.array([1.0]), bijector=tfb.Softplus())
-    name: str = "Periodic"
+    lengthscale: Float[Array, "D"] = param_field(jnp.array([1.0]), bijector=Softplus)
+    variance: Float[Array, "1"] = param_field(jnp.array([1.0]), bijector=Softplus)
+    period: Float[Array, "1"] = param_field(
+        jnp.array([1.0]), bijector=Softplus
+    )  # NOTE: is bijector needed?
 
-    def __call__(self, x: Float[Array, "D"], y: Float[Array, "D"]) -> Float[Array, "1"]:
+    def __call__(self, x: jax.Array, y: jax.Array) -> Array:
         """Evaluate the kernel on a pair of inputs :math:`(x, y)` with length-scale parameter :math:`\\ell` and variance :math:`\\sigma`
 
-        TODO: update docstring
+        TODO: write docstring
 
         .. math::
             k(x, y) = \\sigma^2 \\exp \\Bigg( -0.5 \\sum_{i=1}^{d} \\Bigg)
 
         Args:
-            x (Float[Array, "D"]): The left hand argument of the kernel function's call.
-            y (Float[Array, "D"]): The right hand argument of the kernel function's call
+            x (jax.Array): The left hand argument of the kernel function's call.
+            y (jax.Array): The right hand argument of the kernel function's call
         Returns:
             Float[Array, "1"]: The value of :math:`k(x, y)`
         """
